// src/api/generated/models/RecurringFrequency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Manager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Defines how often the task repeats.
 * @export
 */
export const RecurringFrequency = {
    Daily: 'daily',
    Weekly: 'weekly',
    Monthly: 'monthly'
} as const;
export type RecurringFrequency = typeof RecurringFrequency[keyof typeof RecurringFrequency];


export function instanceOfRecurringFrequency(value: any): boolean {
    for (const key in RecurringFrequency) {
        if (Object.prototype.hasOwnProperty.call(RecurringFrequency, key)) {
            if (RecurringFrequency[key as keyof typeof RecurringFrequency] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RecurringFrequencyFromJSON(json: any): RecurringFrequency {
    return RecurringFrequencyFromJSONTyped(json, false);
}

export function RecurringFrequencyFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecurringFrequency {
    return json as RecurringFrequency;
}

export function RecurringFrequencyToJSON(value?: RecurringFrequency | null): any {
    return value as any;
}

export function RecurringFrequencyToJSONTyped(value: any, ignoreDiscriminator: boolean): RecurringFrequency {
    return value as RecurringFrequency;
}