// src/api/generated/models/RecurringTaskStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Manager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Status of a recurring task. Recurring tasks do not have review.
 * @export
 */
export const RecurringTaskStatus = {
    Todo: 'todo',
    InProgress: 'in_progress',
    Done: 'done'
} as const;
export type RecurringTaskStatus = typeof RecurringTaskStatus[keyof typeof RecurringTaskStatus];


export function instanceOfRecurringTaskStatus(value: any): boolean {
    for (const key in RecurringTaskStatus) {
        if (Object.prototype.hasOwnProperty.call(RecurringTaskStatus, key)) {
            if (RecurringTaskStatus[key as keyof typeof RecurringTaskStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function RecurringTaskStatusFromJSON(json: any): RecurringTaskStatus {
    return RecurringTaskStatusFromJSONTyped(json, false);
}

export function RecurringTaskStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecurringTaskStatus {
    return json as RecurringTaskStatus;
}

export function RecurringTaskStatusToJSON(value?: RecurringTaskStatus | null): any {
    return value as any;
}

export function RecurringTaskStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): RecurringTaskStatus {
    return value as RecurringTaskStatus;
}